requirements.txt
Flask==2.3.3
celery==5.3.1
redis==4.5.5
SQLAlchemy==2.0.20
python-dotenv==1.0.0
email-validator==1.3.1
phonenumbers==8.13.14
requests==2.31.0
twilio==8.4.1